# docs/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ContactsOut": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "live_chat": {
            "title": "Live Chat",
            "type": "boolean"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "telegram": {
            "title": "Telegram",
            "type": "string"
          }
        },
        "required": [
          "telegram",
          "phone",
          "email",
          "live_chat"
        ],
        "title": "ContactsOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "description": "Yagona xato formati (BE-1801). Qo'shimcha maydonlar koddan kelib chiqadi.",
        "properties": {
          "attempts_left": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempts Left"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retry_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "FaqOut": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "id",
          "question",
          "answer"
        ],
        "title": "FaqOut",
        "type": "object"
      },
...
      "InsightAction": {
        "enum": [
          "set_budget",
          "remind_me",
          "review",
          "turn_on_autosave"
        ],
        "title": "InsightAction",
        "type": "string"
      },
      "InsightActionIn": {
        "additionalProperties": false,
        "description": "set_budget: limit (ixtiyoriy); turn_on_autosave: goal_id, amount (ixtiyoriy).",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000000000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "goal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Id"
          },
          "limit": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000000000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          }
        },
        "title": "InsightActionIn",
        "type": "object"
      },
      "InsightActionOut": {
        "properties": {
          "auto_save_monthly": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Save Monthly"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "goal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Id"
          },
          "monthly_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Limit"
          },
          "reminder_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Id"
          },
          "status": {
            "$ref": "#/components/schemas/InsightStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "InsightActionOut",
        "type": "object"
      },
      "InsightKind": {
        "enum": [
          "overspend",
          "discount_days",
          "subscriptions",
          "autosave"
        ],
        "title": "InsightKind",
        "type": "string"
      },
      "InsightListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InsightOut"
            },
            "title": "Items",
            "type": "array"
          },
          "potential_saving": {
            "title": "Potential Saving",
            "type": "integer"
          }
        },
        "required": [
          "potential_saving",
          "items"
        ],
        "title": "InsightListOut",
        "type": "object"
      },
      "InsightOut": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/InsightAction"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/InsightKind"
          },
          "saving": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saving"
          },
          "status": {
            "$ref": "#/components/schemas/InsightStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "icon",
          "category_id",
          "title",
          "body",
          "saving",
          "action",
          "status"
        ],
        "title": "InsightOut",
        "type": "object"
      },
      "InsightStatus": {
        "enum": [
          "active",
          "dismissed",
          "reviewed",
          "done"
        ],
        "title": "InsightStatus",
        "type": "string"
      },
...
      "SupportSessionOut": {
        "properties": {
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "user_hash": {
            "title": "User Hash",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "user_id",
          "user_hash"
        ],
        "title": "SupportSessionOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Finora API v1. Barcha vaqtlar ISO 8601 UTC; mijoz vaqt zonasi `X-Timezone` sarlavhasida. Summalar — so'm (butun son). Xato formati: `{code, message, request_id}` (+ `fields`, `retry_after`, `attempts_left`). Yozuv yaratuvchi POST'lar `Idempotency-Key` talab qiladi. Xato kodlari: docs/backend/ERROR_CODES.md.",
    "title": "Finora API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/help/contacts": {
      "get": {
        "operationId": "contacts_v1_help_contacts_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Contacts",
        "tags": [
          "help"
        ]
      }
    },
    "/v1/help/faq": {
      "get": {
        "operationId": "faq_v1_help_faq_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 16,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FaqOut"
                  },
                  "title": "Response Faq V1 Help Faq Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Faq",
        "tags": [
          "help"
        ]
      }
    },
...
    "/v1/insights": {
      "get": {
        "description": "Qoidaga asoslangan tavsiyalar; matn user tilida (BE-901, BE-902 \"Show dismissed\").",
        "operationId": "list_insights_v1_insights_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_dismissed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Dismissed",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timezone"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "List Insights",
        "tags": [
          "insights"
        ]
      }
    },
    "/v1/insights/{insight_id}/action": {
      "post": {
        "description": "set_budget → kategoriya limiti; remind_me → eslatma; turn_on_autosave → goal\nauto-save; review → reviewed (BE-902).",
        "operationId": "act_v1_insights__insight_id__action_post",
        "parameters": [
          {
            "in": "path",
            "name": "insight_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Insight Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timezone"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/InsightActionIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Act",
        "tags": [
          "insights"
        ]
      }
    },
    "/v1/insights/{insight_id}/dismiss": {
      "post": {
        "operationId": "dismiss_v1_insights__insight_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "insight_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Insight Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Dismiss",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/v1/support/session": {
      "post": {
        "description": "Live chat (BE-1702): chat servisi uchun imzolangan identifikator. Sozlanmagan — 503.",
        "operationId": "support_session_v1_support_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Support Session",
        "tags": [
          "help"
        ]
      }
    },
...
  }
}